summaries: extract getters to dedicated module
